Put node_modules under version control, grudgingly.
This is a "best practice" according to Heroku and this smart fella: http://www.futurealoof.com/posts/nodemodules-in-git.html.  This is because our packages.json only locks in the top-level dependencies.  Lower level dependencies could be different on each deploy, which is clearly a non-starter.  Vendor all the things, bro.